changing from doubles to ints in al files.

+++ EfficientRoutePlanningContraction/app/src/main/java/Project/Graphs/GraphWriter.java
@@ -59,7 +59,7 @@ public class GraphWriter {
                     }
                     writtenEdges.add(edgeKey);
 
-                    double weight = edge.weight();
+                    int weight = edge.weight();
 
                     // Get original node IDs
                     Long nodeId1 = indexToNodeId.get(v1);
